LayoutTests/fast/css/scroll-behavior-exposed-if-disabled.html

<!DOCTYPE html><!-- webkit-test-runner [ CSSOMViewSmoothScrollingEnabled=false ] -->
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests that scroll-behavior is not exposed when the feature is disabled");

shouldBeFalse("'scrollBehavior' in document.documentElement.style");
shouldBeFalse("'scroll-behavior' in getComputedStyle(document.documentElement)");
shouldBeFalse("CSS.supports('scroll-behavior: smooth')");
shouldBeFalse("CSS.supports('scroll-behavior: inherit')");
shouldBeFalse("CSS.supports('scroll-behavior-x: inherit')");
shouldBeFalse("CSS.supports('scroll-behavior-y: inherit')");

</script>
</body>
</html>
