LayoutTests/fast/css/scroll-behavior-exposed-if-disabled-expected.txt

Tests that scroll-behavior is not exposed when the feature is disabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'scrollBehavior' in document.documentElement.style is false
PASS 'scroll-behavior' in getComputedStyle(document.documentElement) is false
PASS CSS.supports('scroll-behavior: smooth') is false
PASS CSS.supports('scroll-behavior: inherit') is false
PASS CSS.supports('scroll-behavior-x: inherit') is false
PASS CSS.supports('scroll-behavior-y: inherit') is false
PASS successfullyParsed is true

TEST COMPLETE
